requirements.txt
# NBA Stats Dashboard Dependencies

# Core
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
click>=8.0.0

# NBA API
nba_api>=1.4.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
joblib>=1.3.0
scipy>=1.10.0

# Optional: Hyperparameter tuning
optuna>=3.0.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Monitoring
sentry-sdk>=2.51.0
